valores trocados para double (voltar para float); conversão de double para string; calculadora trabalhando com double; tela de cálculo de porcentagem e valor juntadas; Criado tela de cálculo de valor futuro;

+++ Budget/Ferramentas/Porcentagem/Porcentagem.swift
@@ -9,46 +9,46 @@
 import Foundation
 
 class Porcentagem : Calculadora {
-    var porcentagem: Float = 0
+    var porcentagem: Double = 0
     
     override init() {
         super.init()
         self.opcao = 4
     }
 
-    func calcularPorcentagemDoValor() -> Float {
+    func calcularPorcentagemDoValor() -> Double {
         self.numeroFinal = porcentagem/100
         
         return self.calcularOperacao()
     }
     
-    func calcularValorComDesconto() -> Float {
+    func calcularValorComDesconto() -> Double {
         self.porcentagem /= 100
         self.numeroFinal = 1 - porcentagem
         return self.calcularOperacao()
     }
     
-    func calcularValorComJuros() -> Float {
+    func calcularValorComJuros() -> Double {
         self.porcentagem /= 100
         self.numeroFinal = 1 + porcentagem
         return self.calcularOperacao()
     }
     
-    func calcularValorPorcentagem() -> Float {
+    func calcularValorPorcentagem() -> Double {
         self.numeroFinal = self.numeroAtual/self.numeroFinal
         self.numeroAtual = 100
         return self.calcularOperacao()
     }
     
-    func calcularAumentoPercentual() -> Float {
+    func calcularAumentoPercentual() -> Double {
         return self.calcularValorPorcentagem() - 100
     }
     
-    func calcularDiminuicaoPercentual() -> Float {
+    func calcularDiminuicaoPercentual() -> Double {
         return 100 - self.calcularValorPorcentagem()
     }
     
-    func calcularValorInicialAumentado() -> Float {
+    func calcularValorInicialAumentado() -> Double {
         //numeroFinal = 1200
         //porcentagem = 20
         self.porcentagem += 100
@@ -57,7 +57,7 @@ class Porcentagem : Calculadora {
         return self.calcularOperacao()
     }
     
-    func calcularValorInicialDiminuido() -> Float {
+    func calcularValorInicialDiminuido() -> Double {
         self.porcentagem = 100 - self.porcentagem
         self.numeroFinal /= self.porcentagem
         self.numeroAtual = 100
